some more dead code removal. now mostly related to WinNT

+++ src/moab/MPBSI.c
@@ -549,7 +549,7 @@ int MPBSWorkloadQuery(
   short  TaskList[MAX_MTASK_PER_JOB + 1];
   char   Message[MAX_MLINE];
 
-  int    OldState;
+  long   OldState;
 
   mjob_t *JNext;
 
@@ -2432,7 +2432,7 @@ int MPBSNodeLoad(
           int UseUtil = FALSE;
           int UseDed  = FALSE;
 
-          int OldState;
+          long OldState;
 
           OldState = N->State;
 
@@ -2887,7 +2887,7 @@ int MPBSNodeUpdate(
   time_t        tmpTime;
 
   int           tmpProcs;
-  int           OldState;
+  long          OldState;
 
   const char *FName = "MPBSNodeUpdate";
 
@@ -3572,7 +3572,7 @@ int MPBSJobUpdate(
   int                  RMIndex)  /* I */
 
   {
-  int           OldState;
+  long          OldState;
 
   char          tmpBuffer[MAX_MBUFFER];
